:recycle: Extract InlineStylingOptions into own component
